Tweak HUD odometer (Pt.2)
Remove code for drawing the right side of odometer frame; the frame pixels are added to the bitmap font. This saves ~44 bytes, a few cycles faster per frame, but removes the night time ‘highlight’ on the odometer.

Remove the extra digit (leading zero) and re-center the graphic. This may save a few cycles per frame but alters the HUD visuals.